extraordinaria_site/backend/requirements.txt
openai
requests
PyJWT
python-dotenv
flask
flask-cors
bcrypt
pandas
plotly
stripe
simple-salesforce
firebase-admin
google-auth
